Use Node's native fetch()
